fix tests/test warnings
This patch fixes the failing it_can_read_write test
by clobbering the environment before invoking bash in
the client so that the user's PS1 will not worm its
way into the output. My guess is that this bug did
not come up before because the author used zsh
rather than bash as their shell.

It also fixes several warnings in the test code.