fix significant unsoundness issues
This patch fixes a bunch of unsoundness issues
in this crate. To do so, we needed to break the
public interface (in particular, we needed to
remove the AsRawFd impl on Slave and Master), so
this is a breaking change. As such this is a minor
version bump (since we are pre-1.0).